Moved database functions to a different file and README update
